requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
pillow==10.3.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
